requirements.txt
asgiref==3.8.1
Django==5.0.4
django-rest-authtoken==2.1.4
djangorestframework==3.15.1
sqlparse==0.4.4
typing_extensions==4.11.0
